Add joi2024yo1a/joi2024_yo1a_a/joi2024_yo1a_a.cpp

+++ joi2024yo1a/joi2024_yo1a_a/joi2024_yo1a_a.cpp
@@ -3,7 +3,7 @@ int main(){
   int x, y;
   std::cin >> x >> y;
   std::cout << x + y + 3 << "\n";
-  for(long long i = 0; i < 999999; ++i){
+  for(int i = 0; i < 999999; ++i){
     std::cerr << "A";
   }
 }
